+++ src/screen.cpp
@@ -64,11 +64,11 @@ clear_summary (Population population)
  *                evaluate_consecutive_weekends below for further information.
  * =====================================================================================
  */
-unsigned long
+unsigned long long
 evaluate_extra_shifts (Chromosome chromo, unsigned worker)
 {
-  unsigned long value = 0;
-  unsigned long power = 0;
+  unsigned long long value = 0;
+  unsigned long long power = 0;
   unsigned together = 0;
   unsigned base = get_shift_base (chromo);
   worker %= chromo.width;
@@ -212,7 +212,7 @@ punish_long_shifts (Chromosome chromo)
   unsigned base = get_shift_base (chromo);
 
   for (unsigned worker = 0; worker < chromo.width; worker++)
-    for (unsigned long extra_shifts =
+    for (unsigned long long extra_shifts =
 	 chromo.summary.extra_shifts[worker], n = 1; extra_shifts > 0;
 	 extra_shifts /= base, n++)
       penalty +=
